fixed other tests
